modify field rating on Review

+++ client/models.py
@@ -25,7 +25,7 @@ class Restaurant(models.Model):
 
 
 class Review(models.Model):
-    rating = models.IntegerField()
+    rating = models.DecimalField(max_digits=2, decimal_places=1)
     text = models.CharField(max_length=254)
     restaurant = models.ForeignKey(Restaurant, on_delete=models.CASCADE)
     author = models.ForeignKey(Customer, on_delete=models.CASCADE)
